separated out types
